common/cnxk: fix build on Ubuntu 24.04
[ upstream commit 20c29a0e4602b9c7be5ea299457f909846c3785d ]

Due to different datatypes, warnings are thrown for writing
on space more than its size.

Bugzilla ID: 1513
Fixes: 39ac394aa7a8 ("common/cnxk: fix device MSI-X greater than default value")

+++ drivers/common/cnxk/roc_irq.c
@@ -15,7 +15,7 @@
 
 #define MSIX_IRQ_SET_BUF_LEN                                                   \
 	(sizeof(struct vfio_irq_set) + sizeof(int) *			       \
-			(plt_intr_max_intr_get(intr_handle)))
+			((uint32_t)plt_intr_max_intr_get(intr_handle)))
 
 static int
 irq_get_info(struct plt_intr_handle *intr_handle)
